(#8414) Create scheduled_task type for use with Windows scheduled tasks
This creates a new type for interacting with scheduled tasks on
Windows with a provider that uses the Win32::TaskScheduler gem.
Currently this type is limited to creating & managing time based
triggers.

The following time based triggers are supported, with any limitations
mentioned below: One-time, daily, weekly, monthly date-based, monthly
day-of-week-based.

  - 'day_of_week' must be specified as one of 'mon', 'tues', 'wed',
    'thrus', 'fri', 'sat', or 'sun'.  Support for cron-style day of
    week numbers (0 = Sunday, ..., 5 = Friday, etc.) are not supported
    at this time.

  - 'months' must be specified as a numeric (1 = January, ..., 12 =
    December).