(Maint) Remove rspec from shebang line
Without this patch Ruby 1.9 is still complaining loudly about trying to
parse the spec files.  The previous attempt to clean up this problem in
edc3ddf works for Ruby 1.8 but not 1.9.

I'd prefer to remove the shebang lines entirely, but doing so will cause
encoding errors in Ruby 1.9.  This patch strives for a happy middle
ground of convincing Ruby it is actually working with Ruby while not
confusing it to think it should exec() to rspec.

This patch is the result of the following command run against the source
tree:

    find spec -type f -print0 | \
      xargs -0 perl -pl -i -e 's,^\#\!\s?/(.*)rspec,\#! /usr/bin/env ruby,'