(PUP-3247) Remove trigger_string method
Previously, our vendored task scheduler code defined a method that if
called would never work. This is because it would try to return the
`trigger` local variable, but it was never defined, and so would call
the `#trigger` method instead, and that requires one argument for the
index.

This commit just removes the method since it's never used in puppet.