(PUP-3912) Revert transpec's use of double
An earlier commit in the rspec 3 port used transpec to make
wholesale changes. However, one of transpec's changes is to
replace 'stub' with 'double' which relies on the rspec-mocks
gem, which I ran into difficulties with.

So this commit reverts those changes. If we sort out issues
with rspec-mocks at a later date, we can move to 'double'
across the board.