(PUP-1291) Support Scheduled Tasks Repeat Triggers
Previously, the scheduled_task resource did not support repetition for
anything less than 1 day. This is limiting as the cron resource can
support repetition in minutes. Task scheduler on Windows can support
repeating tasks every x number of minutes, and the win32-taskscheduler
gem also allows this.

Allow Puppet's scheduled_task resource to support repetition every x
number of minutes. Without this change the schedule resource continues
to be more limited than on other platforms.

Also, behave similarly to the defaults in task scheduler, so that an
end user might be able to easily specify a task that repeats on a 30
minute interval without specifying a duration:

scheduled_task { 'doIt':
  command => 'C:\\Windows\\System32\\notepad.exe',
  trigger => [{ schedule => 'daily',
                minutes_interval => '30',
                start_time => '15:20'}]
}