(PUP-5123) Replace http:// links with https://
This replaces most http:// links with https:// The
replaced ones have all been verified to serve
identical content.  Those which aren't replaced are
contained within XML schema or are internal puppetlabs
links which are unable to be verified externally. Some
are also outdated and inaccessible and should be updated
seperately.