(MODULES-6264) Add Scheduled Task V2 to V1 adapter
This commit adds a new class which has the same interface as the older V1
Scheduled Tasks (puppet/util/windows/taskscheduler.rb).  This adapter class will
implement the V1 methods but using the V2 API calls.  This commit only adds the
minimum required implementations, and the rest with NotImplementedErrors.  These
will be populated in later commits.