(MODULES-6266) Add ScheduleTask Provider for V2 API adapter
This commit adds a non-default provider for the Scheduled Task type but instead
uses the V2 adapter class.  This provider is in essence a verbatim copy of the
original provider (win32_taskscheduler) but changes the helper class to the
V2 API adapter object.  This maintains backwards compatilbity and will allow the
older V1 implementation to be deprecated and then removed easily.