(MODULES-6395) Implement minimal new type
In preparation for determining whether separating scheduled tasks into two types,
this commit copies the existing v2 provider and v1 type to create the new
`sched_task` type, changing as little code as possible to get the type and the
provider to function. Creates the new example files for the new type as well.

Logs for the output from this initial pass can be found here:

+ https://gist.github.com/michaeltlombardi/ee0522c96a02795ce40ffb1e8146d37f