(MODULES-6398) Fix error message on invalid day_of_week trigger
Previously if an invalid value for the day_of_week property was passed (e.g.
MON) it would fail with `Cannot convert "nil" to Fixnum`.  This commit adds
invalid value detection when building the bitmasked value and raises an error
which includes the bad value.  This commit also adds a test for this scenario.