(MODULES-6268) Fix error when switching monthly trigger types
Previously when changing from a monthly day of week to a monthly date trigger
type (and vice versa) Puppet would raise an error.  This was due to Puppet using
the same schedule name (monthly) for two distinct trigger types in windows API.
This commit adds an additional check to the tiggers_same? method which also
evaluates the underlying trigger type, not just the Puppet representation to
determine if two triggers are indeed the same.  This commit also adds tests for
this scenario.