(MODULES-6417) Add usage and examples to README
Prior to this commit the documentation for the `scheduled_task` module
largely pointed back to the official Puppet documentation for information
on the type reference and on the usage of the resource. This commit
ensures that that information now lives in the scheduled_task module and
adds numerous clarifying examples of use.

This will allow future iterative work to continue in this module without
needing to point back to the not-actively-developed type and provider in
core Puppet. Documentation updates to the module can now be safely made.