(MODULES-6526) Confine compatibility to a feature
Prior to this commit any calls to the win32 legacy provider after
the introduction of the compatibility setting cause a run to blow
up because that provider does not have a way to manage compatibility.

This commit adds a feature, `:compatibility`, and limits the methods
for compatibility setting and getting to that feature.  Note that this
method will cause the legacy provider and any provider that does not
support the feature to silently ignore compatibility if passed in a
manifest.  It will not error but it also won't do anything.

If swapping between providers, this can probably cause some strange
behavior.  All net-new features and properties for this module will
probably need to be added this way to prevent explosions in the
legacy provider.

The added acceptance test is a simple copy-change and is not robust.
A future ticket and PR should ensure that the acceptance suite does
a bit more to test against both providers.