(MAINT) Refactor spec to iterate over providers
Prior to this commit the spec for the providers iterated
over the helpers used.  This causes some problems with
plans to copy the new provider into the old as both
providers will temporarily use the same provider.

This commit changes the spec logic to iterate over the
providers and map the appropriate helpers to them. No
tests are changed or affected.