(MAINT) Replace Win32::TaskScheduler constant
Prior to this commit the code relied on the Win32::TaskScheduler class
to provide the DISABLED constant, indicating a task is disabled. This
required loading that class and caused puppet resource calls to fail.

This commit adds the task flag constants to the `TaskScheduler2` class
and bypasses the need to load the Win32::TaskScheduler at all. It then
replaces all calls for that constant to the appropriate call to those
defined in `TaskScheduler2`.