(MAINT) Update Readme with recent improvements
Prior to this commit the readme included only vague
information about the compatibility level property
and language about the providers which is inaccurate
in light of the change to use the V2 API for the
legacy provider.

This commit expands on the compatibility property
documentation, clarifying that it applies only to
the new provider and which compatibility settings
can be set, as well as updating the language for
the providers to clarify that both of them use the
V2 API but that only the new provider will get new
features and improvements.