(MODULES-7120) Refactor triggers_same?
Prior to this commit the logic in triggers_same? required three checks
and several steps to build up intermediate objects to compare. This
commit simplifies the checking logic by canonicalizing the desired
manifest hash so that it will match the formatting of the current
(which is canonicalized from the iTrigger on disk into a manifest hash)
and then relying on a hash comparison. This method ensures that the
current_trigger is always a superset of the key value pairs specified
in the `desired_trigger` and that every matching key has the same
values.

In order for this canonicalization to fully work minor changes had to
be made to `canonicalize_and_validate_manifest` - previously this
method did _not_ return the `start_date` or `start_time` keys with the
correct formatting produced and expected by other methods.  This commit
ensures that it produces strings formatted correctly. It performs
similar conversions with minutes_duration and minutes_interval to
convert string values to numeric values.

This commit also updates the tests; prior to this commit the unit
tests for win32taskscheduler_spec did not use the canonical
representation of a manifest hash for the `current` values being passed
to `trigger_same?`. This commit updates the tests to correct for this
problem.

It also adds some missing basic tests around
canonicalize_and_validate_manifest