(MAINT) Add legacy taskscheduler code as helper and freeze
Prior to this commit the spec tests relied on the built-in
utility code for the windows task scheduler inside of core
puppet. With the Puppet 6 release this code will be removed,
meaning the tests relying on this reference will fail.

This commit copies that code into a helper file in the spec
folder and ensures the tests relatively require it instead
of pointing to the utility baked into Puppet.

This should ensure the tests continue to pass and backwards
compatibility can be safely maintained once Puppet 6 is
released.