(MODULES-7129) Add logon trigger
This commit adds the logon event trigger to the module.

It allows the user to specify in the manifest:

- A local or builtin user
- Either an empty string or `undef`, both of which translate to
  running the task whenever _any_ user logs in.

Whether this works with an active directory account is currently
untested, but it works by converting the given username to a SID
and back - so if the AD account can be resolved it should still
work.

If an unresolvable user is specified the manifest will raise an
error letting the user know why.

This commit adds minimal additional testing to the suite.