(MODULES-7203) Support nonroot task folders
Prior to this commit the type and providers did not support specifying subfolders in which
to place a scheduled task. This commit adds support to the type and the taskscheduler_api2
provider for specifying scheduled tasks in subfolders.

This feature only exists for tasks whose compatibility is 2 or higher which prevents it from
being used on the legacy win32_taskscheduler provider altogether and necessitates some
validation prior to runtime.

This commit adds such validation by failing with descriptive errors during resource validation.

This new functionality requires the ability to scaffold folders in which to place the scheduled
task, which now happens automatically and only if the folder path does not already exist.

This commit does NOT add functionality for pruning the folders on task deletion, this should be
addressed in a future commit.

It does update the documentation and tests for the new feature as well as ensure all existing
tests continue to pass.