Update error message expectations
Encountering errors on AppVeyor as the error message being returned for this negative test is:
```
ERROR: The system cannot find the path specified.
```
whereas the test is expecting the error:
```
ERROR: The system cannot find the file specified.
```

Still need investigation as to why there is a difference in the error message ("file" vs "path"). Unable to reproduce with Windows images provisioned locally (tried on Server 2008 R2 and Server 2016). There may be something unique to AppVeyor's particular images.