(MODULES-10986) Fix gMSA username support
Since the enhancement intended to help sysadmins, pulled in as
puppetlabs/puppetlabs-scheduled_task#150: When not setting a password,
this is interpreted as "Run only when user is logged on".

When configuring a scheduled task to use a Group Managed Service Account
(gMSA) principal, there is no password to provide. The scheduler will
get the password automatically when properly configured, and should be
able to run the task even though the service account is not "logged on".
Otherwise, the Managed Service Account is never "logged on" so the task
does not run on expected schedule.

It does not seem viable to add a proper test case for this particular
situation without an AD domain configuration and a proper gMSA in place.
Without a correct gMSA, an error message stating "No mapping between
account names and security IDs was done" appears to be returned.